+++ Source/Flock/Private/FlockSubsystem.cpp
@@ -8,7 +8,7 @@
 #include "Engine/World.h"
 
 const FString UFlockSubsystem::ApiVersion = TEXT("v1");
-const FString UFlockSubsystem::SdkVersion = TEXT("0.3.0");
+const FString UFlockSubsystem::SdkVersion = TEXT("0.4.0");
 
 UFlockSubsystem* UFlockSubsystem::Get(const UObject* WorldContextObject)
 {